pyproject.toml
[project]
name = "scrapontologies"
version = "0.1.0"
description = "Library for extracting schemas and building ontologies from documents using LLM"
authors = [
    { name = "Lorenzo Padoan", email = "lorenzo.padoan977@gmail.com" },
    { name = "Marco Vinciguerra", email = "mvincig11@gmail.com" },
    { name = "Marco Perini", email = "perinim.98@gmail.com" }
]

dependencies = [
    "certifi>=2024.7.4",
    "charset-normalizer>=3.3.2",
    "idna>=3.8",
    "pillow>=10.4.0",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "urllib3>=2.2.2",
    "langgraph>=0.2.31",
    "psycopg2>=2.9.9",
    "neo4j>= 5.25.0",
    "langchain>=0.3.0",
    "langchain-core>=0.3.10",
    "langchain-openai>=0.2.2",
    "langchain-anthropic>=0.2.3",
    "langchain-google-vertexai>=2.0.5",
    "langchain-google-genai>=2.0.1",
    "langchain-ollama>=0.2.0"
]

license = "MIT"
readme = "README.md"
homepage = "https://scrapegraphai.com/"
repository = "https://github.com/ScrapeGraphAI/scrape_schema"
documentation = ""
keywords = [
    "scrape_schema",
    "ontologies",
    "documents",
    "knowledge graph",
    "scrapegraph",
    "scrapegraphai",
    "langchain",
    "ai",
    "artificial intelligence",
    "gpt",
    "machine learning",
    "rag",
    "nlp",
    "natural language processing",
    "openai",
    "scraping",
    "web scraping",
    "web scraping library",
    "web scraping tool",
    "webscraping",
    "graph",
    "structured data",
    "unstructured data"

]
classifiers = [
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
requires-python = ">=3.9,<4.0"

[project.optional-dependencies]
renderers = ["pyecharts>=2.0.6"]
docs = ["sphinx>=6.0", "furo>=2024.5.6"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-mock>=3.14.0",
    "-e file:.[renderers]",
    "-e file:.[docs]",
    "pylint>=3.2.5",
]

[tool.rye.scripts]
pylint-local = "pylint scrape_schema/**/*.py"
pylint-ci = "pylint --disable=C0114,C0115,C0116 --exit-zero scrape_schema/**/*.py"
update-requirements = "python 'manual deployment/autorequirements.py'"
